構造体を別ファイルに分割
